feat : 일별, 주별, 월별 상품별 예약 건수 조회 (#132)

+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/product/presentation/response/HotProductResponse.java
@@ -12,7 +12,7 @@ public class HotProductResponse {
     private String productName;
     private String productImage;
     private Integer favoriteCount;
-    private Integer reservationCount;
+    private Long reservationCount;
     private Integer price;
     private Integer stock;
     private Boolean isFavorite;
